# [wip] perf(ext/http): improve throughput by ~10k reqs