--- a/serde_v8/magic/v8slice.rs
+++ b/serde_v8/magic/v8slice.rs
@@ -66,12 +66,13 @@ impl V8Slice {
   }
 }
 
+#[inline]
 pub(crate) fn to_ranged_buffer<'s>(
   scope: &mut v8::HandleScope<'s>,
   value: v8::Local<v8::Value>,
 ) -> Result<(v8::Local<'s, v8::ArrayBuffer>, Range<usize>), v8::DataError> {
-  if value.is_array_buffer_view() {
-    let view: v8::Local<v8::ArrayBufferView> = value.try_into()?;
+  let view: Result<v8::Local<v8::ArrayBufferView>, _> = value.try_into();
+  if let Ok(view) = view {
     let (offset, len) = (view.byte_offset(), view.byte_length());
     let buffer = view.buffer(scope).ok_or(v8::DataError::NoData {
       expected: "view to have a buffer",
@@ -85,6 +86,7 @@ pub(crate) fn to_ranged_buffer<'s>(
 }
 
 impl FromV8 for V8Slice {
+  #[inline]
   fn from_v8(
     scope: &mut v8::HandleScope,
     value: v8::Local<v8::Value>,