@lucacasonato 2022-07-01
How does this affect benchmarks which read headers?