@littledivy 2022-07-01
@lucacasonato For `cli/bench/http/deno_http_read_headers.js` -

```
# Deno 1.23.2
Running 20s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.37ms    3.41ms  54.86ms   92.51%
    Req/Sec    12.87k     9.29k   38.83k    77.75%
  Latency Distribution
     50%  230.00us
     75%    0.87ms
     90%    3.58ms
     99%   16.35ms
  512704 requests in 20.07s, 68.94MB read
Requests/sec:  25542.63
Transfer/sec:      3.43MB
```

```
# This patch
Running 20s test @ http://127.0.0.1:4500/
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     4.09ms   12.37ms 136.14ms   92.42%
    Req/Sec    17.63k    14.92k   40.99k    57.79%
  Latency Distribution
     50%  156.00us
     75%    0.87ms
     90%   11.17ms
     99%   66.98ms
  699162 requests in 20.10s, 94.01MB read
Requests/sec:  34780.46
Transfer/sec:      4.68MB
```